+++ src/main/java/org/oagi/srt/gateway/http/api/context_management/data/ContextCategory.java
@@ -3,12 +3,13 @@ package org.oagi.srt.gateway.http.api.context_management.data;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.util.Date;
 
 @Data
 public class ContextCategory implements Serializable {
 
-    private long ctxCategoryId;
+    private BigInteger ctxCategoryId;
     private String guid;
     private String name;
     private String description;
